requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
requests==2.31.0
google-api-python-client==2.133.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
